+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.2.10
 - Another small variant (mini - 6x6)
 - *buildRandomPosition()* for generating arbitrary random positions, see Variant.miniRandom for an example
+- Various minor improvements
 
 ### 0.2.9
 - Fixed FEN validation for small boards